@@ -3,7 +3,7 @@
   "description": "Auto-fixable sort rules for ESLint.",
   "version": "0.0.0-semantically-released",
   "author": "Mark Skelton",
-  "packageManager": "yarn@3.2.1",
+  "packageManager": "pnpm@7.29.1",
   "repository": "github:mskelton/eslint-plugin-sort",
   "homepage": "https://github.com/mskelton/eslint-plugin-sort#readme",
   "bugs": {
@@ -16,42 +16,52 @@
     "sort",
     "fix"
   ],
-  "main": "lib/index.js",
-  "files": [
-    "lib"
-  ],
   "scripts": {
-    "build": "babel src --out-dir lib --extensions .ts --env-name=build",
+    "build": "tsup ./src/index.ts --format cjs,esm",
     "format": "prettier --write .",
     "lint": "eslint . --ext .ts",
-    "test": "jest",
+    "test": "NODE_OPTIONS=--experimental-vm-modules pnpm jest",
     "ts": "tsc"
   },
+  "type": "module",
+  "exports": {
+    "import": "./dist/index.js",
+    "require": "./index.cjs"
+  },
+  "files": [
+    "dist",
+    "index.cjs"
+  ],
   "peerDependencies": {
     "eslint": ">=6"
   },
   "dependencies": {
-    "@typescript-eslint/experimental-utils": "^5.27.1",
+    "@typescript-eslint/experimental-utils": "^5.54.1",
+    "isomorphic-resolve": "^1.0.0",
     "natural-compare": "^1.4.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.17.10",
-    "@babel/core": "^7.18.2",
-    "@babel/preset-env": "^7.18.2",
-    "@babel/preset-typescript": "^7.17.12",
-    "@mskelton/tsconfig": "^1.1.1",
-    "@types/eslint": "^8.4.3",
-    "@types/jest": "^28.1.1",
-    "@types/natural-compare": "^1.4.0",
-    "@types/node": "^17.0.41",
-    "@typescript-eslint/eslint-plugin": "^5.27.1",
-    "@typescript-eslint/parser": "^5.27.1",
-    "babel-jest": "^28.1.1",
-    "eslint": "^8.17.0",
-    "eslint-config-prettier": "^8.5.0",
-    "jest": "^28.1.1",
-    "prettier": "^2.6.2",
-    "semantic-release": "^19.0.2",
-    "typescript": "^4.7.3"
+    "@babel/cli": "^7.21.0",
+    "@babel/core": "^7.21.0",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-typescript": "^7.21.0",
+    "@jest/globals": "^29.5.0",
+    "@mskelton/tsconfig": "^2.0.0",
+    "@types/eslint": "^8.21.1",
+    "@types/estree": "^1.0.0",
+    "@types/jest": "^29.4.0",
+    "@types/natural-compare": "^1.4.1",
+    "@types/node": "^18.15.0",
+    "@typescript-eslint/eslint-plugin": "^5.54.1",
+    "@typescript-eslint/parser": "^5.54.1",
+    "babel-jest": "^29.5.0",
+    "eslint": "^8.36.0",
+    "eslint-config-prettier": "^8.7.0",
+    "jest": "^29.5.0",
+    "prettier": "^2.8.4",
+    "semantic-release": "^20.1.1",
+    "ts-jest": "^29.0.5",
+    "tsup": "^6.6.3",
+    "typescript": "^4.9.5"
   }
 }
